Update patch set 1
Patch Set 1:

> I think looking for this by section would be better, as if you take
 > Ian's suggestion to not have the buildid note mapped, there won't
 > (AIUI) be a PT_NOTE for it. Does readELFNote not work for some
 > reason?

I am very sensitive to minimizing I/O. It has a direct impact on go command latency, the time between when the go command starts and when it starts doing useful work.

readELFNote reads lots more data from the file: it needs to load the section headers and then the section string table in order to search by name. That seems unavoidable for shared libraries (and in general I worry less about the speed of that mode), but it's easy to avoid for binaries, and I have. (Note that readELFGoBuildID is explicitly making sure that debug/elf doesn't do anything with sections.)

Patch-set: 1
Label: Code-Review=0