cmd/internal/obj, cmd/link: change object files to have a symbol table
DO NOT SUBMIT

I had some time for some recreational hacking and I'd been wondering how much
using a symbol table in object files would speed up the linker. It turns out
that it was a lot more than I expected: this CL improves all.bash by about
25-30s on my laptop (~6.5 mins to ~6 mins). So even though it's late in the
release to do this, is that enough of an improvement to be worth it?

This is mostly just the simplest thing I could do, I didn't think about things
very hard.

Updates #10571

Change-Id: I24eb071425c77b1c611d8f1f4e63deee22887e76