Update patch set 1
Patch Set 1:

> This does not fix the bug. Memory barriers is always a game of two:
 > a writer must do #StoreStore-style memory barrier between the
 > stores, and a reader must do #LoadLoad-style memory barrier between
 > the loads. While a reader can load values in the wrong order, it
 > does not matter how writer stores them.

I didn't think this would fix the bug. But I'm worried about the performance cost of putting barriers on the two loads done for every pointer in the GC. The stores are relatively infrequent. If we can do something more expensive there, that would be preferable.

Any suggestions on how to proceed?

Patch-set: 1