Update patch set 2
Patch Set 2: Code-Review+1

This solves the problem from a user perspective but does also send needless queries that just get ignored.  I feel this is the simplest way to solve it, but wondering if it wouldn't be better to split suffix trying into its own function and only use it if first attempt returns NXDOMAIN for all types tried.  Maybe something to think about in the future, but for scope of 1.5 I'm ok with this.  Nice cleanup/combination of A/AAAA handling.

Patch-set: 2
Label: Code-Review=+1