Update patch set 2
Patch Set 2:

> Well, we already generate the zero value of basic types. I'd just
 > explicitly handle slices and maps, and maybe just think through the
 > rest of the cases to see whether anything else is clearly missing.

Firing recursive slices is a little trickier, sine there will be a competition between expansion by a factor 'complexSize' and termination with probability '1/complexSize'.

For instance, the following still overflows the stack:

  type R struct {
      N []*R
  }

I have a fix locally that shrinks 'complexSize' as a function of recursion depth. It seems to do the trick, but should perhaps wait for 1.6?

Should I open a new ticket or should we keep all this in #8818?

Patch-set: 2