Update patch set 4
Patch Set 4:

I agree with Matthew's point about multiple suffixes but don't feel this is a clean way to solve it.  It can potentially lead to more useless queries that don't get used in the real world since IPv6 isn't too widely used.  That is, if one has 3 items in search path, this means 4 AAAA queries for each domain resolution though none of these would matter.  I feel we need a way to iterate through the search list one at a time and stop once either returns an A or AAAA.  Wondering if it may be cleaner to have lookupDNS do this by taking a ... of qtypes?

Patch-set: 4
Label: Code-Review=0