Update patch set 6
Patch Set 6:

Thanks Mikio.  Thinking from a troubleshooting perspective, one valuable tool is a packet capture, and it gets real confusing when you see many extra queries being sent but not used, so this is my line of thinking.  The breaking apart as you've done is good, but two general comments.  First, this CL is too big I think, the awesome cleanup you've done in other areas maybe should be in another CL.  I keep making this mistake myself and now understand why others balk at it - it's harder to follow when you didn't write it.  Secondly, I don't think this solves the suffix issue.  Suffix searching should be steppable in a way.  Not sure how to best approach, either by moving steps to the query calling function or using some kind of global or otherwise yieldable nextSuffix field.  For the A/AAAA query, both queries should be sent at the same time for the same suffix.  Once either returns an address, suffixing should stop for both types.  I do not see that this addresses that...but it is late here so excuse if it does this.

Patch-set: 6