Update patch set 6
Patch Set 6:

It should probably be 3 CLs but I'll review one massive one in detail when done if that's what is decided.  1 to clean up (the combination of A/AAAA parsing, splitting the semaphore, etc), 2 to fix any race conditions, and 3 to fix the suffixing issue.  Thanks for tackling any of the three, if we move to Go DNS by default I have a bit of a vested interest in making sure it is done correctly.

Patch-set: 6
Label: Code-Review=0