Update patch set 9
Patch Set 9:

Sorry for taking ages to get back to this. It's subtle and I kept on getting it wrong. The patch I just uploaded works in that it passes the tests it adds and can be used to build juju and its tests.  It doesn't yet cover the case where you say 'go install -linkshared -buildmode=shared pkg' and pkg has dependencies that are not already in shared libraries.

Patch-set: 9