cmd/vet: check for durations without units
DO NOT SUBMIT (will have to wait for Go 1.6)

Check for small time.Duration values specified
as literal integers. These are likely meant
to be in units of seconds, but are not.

This check could be strengthened considerably
in several directions, for example by insisting
on units for all durations, regardless of length.
Data from a corpus suggests that while this would
improve code readability (time.Second is clearer
than 1e9) it would not catch many bugs.

A typical bug is:

var updateInterval = flag.Duration("interval", 10, "The number of seconds between each poll to etcd.")

Of the 6 instances flagged in the standard library,
none were bugs, 5 were in the time package tests,
1 was in the flag package tests,
and all would benefit from an added duration.

Running this check over a large open source corpus
flagged 702 instances:

https://gist.github.com/josharian/9341f27cfd5cc59b2813

A small, random sampling suggested that these were
~15% bugs, ~60% non-bugs but would be improved with units,
and about ~15% stet.

Fixes #11137.

Change-Id: Ibc2558e55351eeab2f84a963f906fe0a8d6136f6