cmd/internal/obj, cmd/link: change object file format to be section oriented
DO NOT SUBMIT -- still just a hack, but it seems an interesting one.

The motivation for this is performance. Linking juju seems to be the test for
this currently, and it improves that by about 21% (about 35% if you pass -w).

Fails all.bash in the gdb runtime test which is a bit odd, disabled for now.

Change-Id: I0171085eed3ec244a41a0bfa4caa5c6fdbcbc517