cmd/internal/obj, cmd/link: change object file format to be section oriented
DO NOT SUBMIT -- still just a hack, but it seems an interesting one.

The motivation for this is performance. Linking juju seems to be the test for
this currently, and it improves that by about 21% (about 35% if you pass -w).

I went a bit overboard on testing and analysis, you can see the numbers here:

https://docs.google.com/spreadsheets/d/14JEb5FM2maeTX70Tada8AUGN1j1NZQ7AqT5iJE50LmU/edit?usp=sharing

(all times are elapsed times for linking the github.com/juju/juju/cmd/juju
binary).

Fails all.bash in the gdb runtime test which is a bit odd, will look at that
tomorrow.

Change-Id: I0171085eed3ec244a41a0bfa4caa5c6fdbcbc517