Update patch set 1
Patch Set 1:

> Do we need to do something similar for windows syscalls?

If so, it is unrelated to this problem. This problem arises because we switched from C wrappers in Go 1.4 to Go wrappers in Go 1.5. This CL reestablishes in Go the semantics that the old C code had. There was not a similar change to Windows system calls. So if they were working in 1.4, they should still be working today.

Patch-set: 1