Update patch set 1
Patch Set 1:

> Thanks Russ. From my cursory look, I'm not 100% convinced that
 > adding an import map is necessary. Maybe the GOPATH mechanism could
 > be reused?
 > 
 > > "For the record, gb does nothing more than GOPATH=$PROJECT/src:$PROJECT/vendor/src
 > go build."
 > 
 > This code accomplishes more though, such as disallowing direct
 > imports from the vendor folder, and enabling compiles even without
 > any GOPATH at all.

The tools use vendor/src not because it is the right design but because it is the design that kind of works without modifications to the go command. We are modifying the go command, so now the design space is less constrained. There's very little point to the src directory except this backwards compatibility. We could leave it but then it would just be a wart we have to explain forever ("why is there an unnecesary extra level of directory under vendor?" would become a FAQ). Better to get the design right and suffer a minor period of incompatibility.

Patch-set: 1
Label: Code-Review=0