Update patch set 3
Patch Set 3:

PTAL (or a first one)

I removed the -vendor flag in favor of an environment variable. The discussion on the list and on HN made it clear that an environment variable would be less invasive for people to use, and there's no point in having two ways.

There are now tests that imports resolve as expected and a test that compilation works.

Patch-set: 3
Label: Code-Review=0