Update patch set 3
Patch Set 3:

(3 comments)

> > I propose we tie this to the vendoring experiment, and do it only
 > > in 'vendor' mode. Then we can revisit this when we consider
 > whether
 > > to lock in the vendor semantics for Go 1.6.
 > 
 > Where is the vendoring experiment code?

It's in https://go-review.googlesource.com/#/c/10923/, not yet submitted.
Once that CL is submitted, please update this one to avoid the submodule commands if go15VendorExperiment == false. The right way to do this is probably to modify the loops themselves, not the global tables. Something like

    if !go15VendorExperiment && strings.Contains(cmd, "submodule") {
        continue
    }

inside each of the for loops is good: it's easy to pull out later, and it works even if go15VendorExperiment is changing from run to run.

Patch-set: 3