Update patch set 1
Patch Set 1:

> This feels too special purpose. If you want to do just one
 > download, why not invoke the git command directly?

Because the logic mapping import path to VCS command is really quite involved. For example, if it's a vanity import path, it's not clear what git command to run or even if it *is* a git command.

I'd like to use this as part of another tool, not just as an occasional one-off.

Patch-set: 1