Update patch set 2
Patch Set 2:

> Why? What benefit is there? Code using this will need a build tag,
 > and code not using this would be just as fast as far as I can tell.
 > 
 > Also the code is wrong (CastagnoliTable = makeTable(IEEE)).

Fixed the embarrassing incorrect code; sorry about that. The tests didn't catch
it on my amd64 machine, because the call to the fast updateCastagnoli in Update
discards the table itself.

I'm not sure I understand your comment about needing a build tag. Do you mean
because this will only be available in Go 1.5 and above? Surely this is true of
every feature that is added to the standard library over time?

The answer to "why" is so that the user doesn't need to do this themselves in every package that wants to compute a crc32c value:

    https://github.com/jacobsa/gcloud/blob/54d39dbdab36b1c1495282bdbd0fae02fa4cd1fc/gcs/gcsfake/bucket.go#L36
    https://github.com/jacobsa/gcloud/blob/54d39dbdab36b1c1495282bdbd0fae02fa4cd1fc/gcs/gcsutil/crc32c.go#L19

hash/adler32 and hash/fnv let the user compute hashes with a single function
call, without appearing to allocate anything per call. So does hash/crc32, but
only for the IEEE table. In contrast, to compute a crc32c (the most common in
my work), I have to either manually set up a package-level variable for each
package that does it, or call MakeTable for each hash, which is (or at least
appears to be) inefficient.

Patch-set: 2
Reviewer: Gerrit User 6475 <6475@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0