Update patch set 6
Patch Set 6:

This results in a test failure if building in a path that is reached via a symlink.  eg my /home is linked to /store2/home and this returns the error below.  If I copy the go source to /tmp, the test passes fine.

--- FAIL: TestVendorRun (0.00s)
	go_test.go:238: running testgo [run hello.go]
	go_test.go:257: standard error:
	go_test.go:258: invalid vendoredImportPath: dir="/store2/home/samm/gomaster/src/cmd/go/testdata/src/vend/hello" root="/home/samm/gomaster/src/cmd/go/testdata" separator="/"
		
	go_test.go:267: go [run hello.go] failed unexpectedly: exit status 1

Patch-set: 6
Reviewer: Gerrit User 7480 <7480@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0