fmt: handle negative width/prec when supplied as an argument
This change makes negative width arguments left align the way a
minus-width in the format string aligns.  The exception is that if the
format string width is already left aligned, then a negative argument
width inverts and goes back to right alignment.

Precision is modified to include an error if the argument is negative.
This differs from a negative precision in a format string which just
terminates the format.

Additional checks for large magnitude widths and precisions are added to
make the runtime behavior (failure, but with different error messages),
more consistent between format string specified width/precision and
argument specified width/precision.

Fixes #11376

Change-Id: I8c7ed21088e9c18128a45d4c487c5ab9fafd13ef