Update patch set 3
Patch Set 3:

In the particular case of nilcheckelim, I wonder whether there's a simpler, clearer solution along the following lines.

Traverse the dominator tree, maintaining a count per control value. When traversing down the tree, increment the control values you encounter. When traversing back up the tree, decrement. When processing a block, if the count for its control value is > 0, then it has been encountered higher in the tree and can be eliminated.

The current dominator tree representation is not convenient for doing a top-down traversal, but it could presumably be cheaply transformed. Thoughts?

We also need a fix for the construction of the dominator tree and its use in the cse pass.  Maybe the trick David describes would be helpful in those areas.

See also CL 11520 for some numbers.

Patch-set: 3