Update patch set 3
Patch Set 3:

There was some further discussion over email. Copying it to gerrit here...

Alan:

Sounds like Vitek-Horspool numbering; is that what you were thinking of?

David:

I was thinking of what we did in 1990 for SSA-of-linked-data and for Modula-3 type tests.
( http://dl.acm.org/citation.cfm?id=93585 -- paper has a monotonicity bug in it, sadly. )
I think that predated Vitek and Horspool, but we never got around to publishing it after
we realized that Wegbreit had figured out the basic algorithm first.

We also didn't do thread-safe update, and we didn't do multiple inheritance, but it's good for static trees.
Friend of ours said that it was a useful hack for representing symbol tables in an attribute grammar, too.
Our plan, back when I cared about Java multiple inheritance of interfaces, was to do optimized dispatch
with graph coloring, using the rule that all the interfaces implemented by class C must have different
colors (i.e., that they are a clique).

Patch-set: 3