Update patch set 1
Patch Set 1: Code-Review-2

R=close

Let's hold off on this. I don't see #11352 ever happening in real code, and this slows down normal code. If someone writes code like that and gets out-of-registers, they will figure out what to do just fine.

Also, at this point in the release cycle, the chance of introducing a semantic or performance bug, thereby hurting real programmers, seems to me higher than the chance of helping any real programmers.

If this really needed to be fixed, we could do at the beginning of the 1.6 cycle. 
But the new SSA backend will fix all bugs along these lines, so I'm inclined to just wait for that.

Patch-set: 1
Label: Code-Review=-2