Update patch set 4
Patch Set 4:

(1 comment)

R=rsc

Unless Adam objects, it sounds like the idea is fine, and as long we can get a backwards-compatible implementation, I think everything is okay. I apologize for the terrible review latency here, but if the revisions can be done in the next week or so I believe this can go into Go 1.5.

Patch-set: 4
Label: Code-Review=0