Update patch set 2
Patch Set 2:

> I'm not sure I understand your comment about needing a build tag.
 > Do you mean
 > because this will only be available in Go 1.5 and above? Surely
 > this is true of
 > every feature that is added to the standard library over time?

It is, yes, and so one requirement for adding a new feature is that it provide a real benefit to outweigh that cost. The benefit here is minimal: the second MakeTable(Castagnoli) runs in basically no time (I clocked it at 10ns), and it is trivial for people to write 'var crc32c = crc32.MakeTable(Castagnoli)' as a global variable if they truly cannot afford 10ns at the call site.

Patch-set: 2