Update patch set 1
Patch Set 1: Code-Review-2

R=close

I see a very minor performance improvement (looks like roughly 0.5% on average for ExpNN and Exp3Power), and a strong security argument against, so that sounds like a no to me. (The security argument also makes the RSA improvements moot.)

As much as possible, it is good to NOT present two different forms of a computation, one for speed and one for security. People will pick the wrong one, or you will be using a library that chose one but you want it to use the other, and so on. It just leads to more complexity. Certainly not for just 0.5%.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2