Update patch set 1
Patch Set 1:

R=close

Please wait until Go 1.6 for this. I'd really like to see this fixed, but it's a bit too late in the cycle to be changing something so subtle. Please add a test demonstrating what the code change is doing, and then remail the change once Go 1.5 has been released. Thanks very much.

Patch-set: 1