Update patch set 12
Patch Set 12: Code-Review+2

I am embarrassed that I dropped this on the floor for so long. My apologies.

It is very late in the Go 1.5 cycle, but I think just barely not too late to get this in. It looks good to me as is. The minor comment Minux made can wait for the next round, and I'm convinced by the argument that these are heap allocated and that stack copying wouldn't change alignment anyway. I am also convinced by the argument in the CL description that the scenario being protected with the allocations is racy.

The most compelling argument for including this in Go 1.5 this late in the game is that it adds these checks for misuse of WaitGroups, where an initial positive Add is called concurrently with Wait. During limited testing of Go 1.5 inside Google I have already seen code that broke due to exactly this misuse of WaitGroup combined with the changes in which goroutines the scheduler runs and the larger default GOMAXPROCS setting. Having this check will make those failures more obviously the fault of a buggy program and not the runtime or sync packages.

Patch-set: 12
Label: Code-Review=+2