Update patch set 1
Patch Set 1:

There's also this text in the documentation:

	For compound operands such as slices and structs, the format
	applies to the elements of each operand, recursively, not to the
	operand as a whole. Thus %q will quote each element of a slice
	of strings, and %6.2f will control formatting for each element
	of a floating-point array.


Also, one expects strings and byte slices to behave the same.

So: which rule applies? Does %x on a byte slice apply to the bytes
or to the slice? You are claiming that the second rule overrides the
first if it is a byte slice printed with %s %q %x. That's arguably right,
but I want confirmation.

Patch-set: 1
Label: Code-Review=0