image/png: add support for Truecolor transparency
This commit adds support for Truecolor transparency in the image/png
library. There is support for both 8 and 16-bit depths (cbTC8/cbTC16).
It will store the "simple transparency" as documented in the
specification [0] as the only value in the palette. The color is checked
against during readImagePass and is set to full transparency if the
current pixel matches.

This also changes the logic in parsetRNS as there can now we a tRNS
chunk without a PLTE chunk, which is the case for Truecolor (color type
2). It is now checked that we are past the header, have not hit the IDAT
chunk, and multiple tRNS chunks have not been seen.

[0] http://www.libpng.org/pub/png/spec/1.2/PNG-Chunks.html

Change-Id: I73c87b363c6f8971b9918b420cec81b398d127fd