cmd/compile: initialize line number properly for temporaries
The expansion of structure, array, slice, and map literals
does not use the right line number in its introduced assignments
to temporaries, which leads to incorrect line number attribution
for expressions in those literals.

This was revealed in CL 9721 because a now-avoided temporary
assignment introduced the correct line number.
I.e. before CL 9721
  "tmp_wrongline := expr"
was transformed to
  "tmp_rightline := expr; tmp_wrongline := tmp_rightline"

This fix specifically targets OAS nodes because the change
in behavior introduced in 9721 is only triggered for OAS nodes.
Other nodes may (still) be wrong, but they are the same wrong
that they were (and for which there appear to be no test cases).
The two relevant changed sections in 9721 (order.go) are:
	case OAS:
		t := marktemp(order)
		orderexpr(&n.Left, order, nil)
		orderexpr(&n.Right, order, n.Left)
		...
and
		if lhs == nil || lhs.Op != ONAME || flag_race != 0 {
			n = ordercopyexpr(n, n.Type, order, 0)
		}
The first change is the ony place a non-nil 3rd arg is
passed to orderexpr, and the second is the only place that
the 3rd param (lhs) is used within orderexpr.

Also includes a repair to CL 10334 involving line numbers
where a spurious -1 remained (should have been 0, now is 0).

Fixes #11400.

Change-Id: I3a4687efe463977fa1e2c996606f4d91aaf22722