Update patch set 1
Patch Set 1:

This isn't right. There are other nodes being created; why should OAS be special? Also, it's never correct to blindly copy a line number from a node without checking Op. setlineno has the canonical logic for that; better to use it than duplicate it.

So I suggest this should be, in the places you've identified,

setlineno(r)

and then the original Nod calls.

Patch-set: 1