encoding/xml: illustrate behaviour change
NOT FOR SUBMISSION

Nigel Tao in https://go-review.googlesource.com/#/c/11635
suggested that it would be good to illustrate the behaviour change
by showing the behaviour of the new tests with the old code.

This PR illustrates that by reverting marshal.go to its previous
contents and changing the new tests so they pass.

Note that a couple of tests are now marked as MarshalOnly because
they don't round trip with the previous code.

Change-Id: I3c494579a3fd38127ccfbd61fbb29b917e8574fd