Update patch set 1
Patch Set 1:

R=go1.6

I think we should reconsider for Go 1.6. The more Types that compare equal pointerwise, the more efficient the SSA common subexpression elimination pass will be. If we could get pointer equality to *be* Type equality and get rid of Eqtype, that'd be even better although I haven't thought through whether this is feasible.

Patch-set: 1