Update patch set 2
Patch Set 2:

I'm not willing to accept that regexp, which boils down to (.*):(.*), as secure.

If we must understand that git origins of that form are secure, I think the thing to do would be to change the code that invokes 'git config remote.origin.url' to turn a non-URL of this form into a URL before returning it back up to the rest of the processing.

I'm sorry for being pedantic about this, but one of the benefits of URLs is that they are so explicit about the meaning. ssh://git@github.com/path is well-defined in a way that git@github.com:path is not. If we start accepting the latter in general, I am afraid that it will lead to a place where can't distinguish different forms correctly anymore. At least if the processing is done when invoking git config remote.origin.url, that quandary is left to a very specific routine and not general processing.

Patch-set: 2
Label: Code-Review=0