Update patch set 8
Patch Set 8:

> Make a suggestion. I still don't know, without digging deeper,
 > whether a server request is an object on the server, and object
 > being sent to the server, or a general description for the request
 > itself.

As far as I can tell, the package consistently uses "server request" to mean an
object on the server -- that is, a request seen by a handler function that is
responsible for generating a response. It uses "client requests" to mean an
object on the client -- a request to be fed to http.Do in order to await a
response from a server.

I agree those phrases aren't ideal, but I'm unable to come up with anything
precise and unambiguous that sounds better to me.

Patch-set: 8