Update patch set 3
Patch Set 3:

> Are these failures okay?

The first one looks bad, the others are ok (preexisting failure and a timeout).

I'm inclined to have you submit anyway.  That first failure has to do with the deadcode changes Josh was working on.  I'm very confused how that affects just Windows, but in any case I don't think it is a bug in your code.

Patch-set: 3