Update patch set 1
Patch Set 1: Code-Review-2

This isn't the bug.  It is fine for values to be args of several phis, or even the same phi.

The bug is in the scheduler.  Regalloc requires that all phis are scheduled before other ops, but the scheduler is not enforcing that.  I'll send a change to fix.

Patch-set: 1
Label: Code-Review=-2