Update patch set 3
Patch Set 3:

Further observation on performance.  If you build the dominator tree (children and siblings recorded), you can do this particular anaysis with a treewalk.  State is an array of boolean indexed by value, per-node state is whether that node was the outermost nilcheck for V, any nilcheck for V when state[V.ID] is true is redundant.

One thing that has started to mystify me a bit is where is the ssa-style symbolic analysis -- i.e., if p1 = phi(p2, p3) and p2 and p3 are nilchecked, then so is p1, and I don't see that here at all.  That requires more than a dominator-tree walk.

We could also, in the forward walk, record all variables whose definitions ensure that they cannot be nil, and use that as well.

Patch-set: 3