Update patch set 4
Patch Set 4:

R=go1.6

I apologize, but this won't make it into Go 1.5. We dropped the ball on this and did not get you a timely review. We're going to work to make sure that doesn't happen for the next cycle.

Patch-set: 4
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0